+++ src/syntax/nodes/ast/node_func.c
@@ -36,7 +36,7 @@ void parse_func(Tokenizer* tokenizer, ASTNode* ast_node, CompilerState* state) {
         Token* parameter_name_token =
             peek_consume(tokenizer, SYMBOL_IDENTIFIER);
         Token* comma =
-            peek(tokenizer, SYMBOL_COMMA) ? consume(tokenizer) : nullptr;
+            (int)peek(tokenizer, SYMBOL_COMMA) ? consume(tokenizer) : nullptr;
 
         ASTNode* parameter_node = create_ast_node(
             AST_IDENTIFIER_FUNCTION_PARAMETER, parameter_name_token);
